fix: - Implement a seperate providers - Exporting `Approvals` and `SelectedIdentity` @types - Implement a seperate `injectedScript` - Implement a seperate event service - Resolving webpack decleration path resolver for the @src/**
