+++ package.json
@@ -113,6 +113,8 @@
     "ts-jest": "^29.1.0",
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
+    "ts-transform-paths": "^3.0.0",
+    "tsconfig-paths-webpack-plugin": "^4.0.1",
     "wait-for-expect": "^3.0.2",
     "webpack": "^5.82.1",
     "webpack-cli": "^5.1.1",
