+++ src/background/services/event/types.ts
@@ -1,8 +1,3 @@
-export interface Approvals {
-  isApproved: boolean;
-  canSkipApprove: boolean;
-}
-
 export type EventHandler = (data: unknown) => void;
 export type EventName = "login" | "identityChanged" | "logout";
 export type Events = {
