+++ src/types/approval/index.ts
@@ -1,0 +1,4 @@
+export interface Approvals {
+    isApproved: boolean;
+    canSkipApprove: boolean;
+}
