+++ src/types/duck/identities.ts
@@ -1,0 +1,4 @@
+export interface SelectedIdentity {
+    commitment: string;
+    web2Provider?: string;
+}
